organizing the archi

+++ microservices/second-service/src/main/java/com/miniprojet/secondservice/services/RendezvousServiceImpl.java
@@ -1,16 +1,26 @@
 package com.miniprojet.secondservice.services;
 
+import com.miniprojet.secondservice.client.FirstServiceRestClient;
+import com.miniprojet.secondservice.entities.Consultation;
 import com.miniprojet.secondservice.entities.RendezVous;
+import com.miniprojet.secondservice.model.Medecin;
+import com.miniprojet.secondservice.model.Patient;
+import com.miniprojet.secondservice.repositories.ConsultationRepository;
 import com.miniprojet.secondservice.repositories.RendezvousRepository;
 import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
-@AllArgsConstructor
+@RequiredArgsConstructor
 public  class RendezvousServiceImpl  implements RendezvousService{
     private RendezvousRepository rendezvousRepository;
+    private RendezvousService rendezvousService;
+    private ConsultationRepository consultationRepository;
+    private FirstServiceRestClient firstServiceRestClient;
 
     @Override
     public RendezVous saveRendezVous(RendezVous rendezVous){
@@ -22,11 +32,23 @@ public  class RendezvousServiceImpl  implements RendezvousService{
     };
     @Override
     public RendezVous getRendezVous(Long rendezvousId) {
-        return rendezvousRepository.findById(rendezvousId).orElse(null);
+        RendezVous rendezVous = rendezvousRepository.findById(rendezvousId).orElse(null);
+        Medecin medecin = firstServiceRestClient.getMedecinById(rendezVous.getMedecinId());
+        Patient patient = firstServiceRestClient.getPatientById(rendezVous.getPatientId());
+        rendezVous.setMedecin(medecin);
+        rendezVous.setPatient(patient);
+        return rendezVous;
     };
     @Override
-    public void DeleteRendezVous(Long id){
-        rendezvousRepository.deleteById(id);
+    public void deleteRendezVous(Long id){
+        RendezVous rendezVous = rendezvousRepository.findById(id).orElse(null);
+        if (rendezVous != null) {
+            Consultation consultation = rendezVous.getConsultation();
+            if (consultation != null) {
+                consultationRepository.deleteById(consultation.getId());
+            }
+            rendezvousRepository.deleteById(id);
+        }
     };
     @Override
     public  RendezVous updateRendezVous(RendezVous rendezVous){
